Reorganize Ledger.Basics in Ledger.HD and Ledger.Abstract
