+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Ledger/Abstract.hs
@@ -1,15 +1,18 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE StandaloneKindSignatures #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
 -- | Interface to the ledger layer
 --
--- This module defines how to apply blocks to a ledger state, and re-exports
--- (from "Ouroboros.Consensus.Ledger.Basics") how to tick ledger states. These
--- are the two main operations we can do with a 'LedgerState'.
+-- This module defines how to tick a ledger state and how to apply blocks to a
+-- ledger state. These are the two main operations we can do with a
+-- 'LedgerState'.
 --
 -- The on-disk values flow as explicit, opaque @Values blk@ arguments (see
 -- 'Ouroboros.Consensus.Ledger.Basics.BlockSupportsLedgerHD'):
@@ -25,6 +28,11 @@
 module Ouroboros.Consensus.Ledger.Abstract
   ( -- * Type-level validation marker
     Validated
+
+    -- * Tick a ledger state
+  , IsLedger (..)
+  , applyChainTick
+  , LedgerError
 
     -- * Apply block
   , ApplyBlock (..)
@@ -50,14 +58,18 @@ module Ouroboros.Consensus.Ledger.Abstract
 
     -- * Re-exports
   , module Ouroboros.Consensus.Ledger.Basics
+  , module Ouroboros.Consensus.Ledger.HD
   ) where
 
 import Control.Monad.Except
 import qualified Control.State.Transition.Extended as STS
-import Data.Kind (Type)
+import Data.Kind
+import GHC.Generics
 import GHC.Stack (HasCallStack)
+import NoThunks.Class
 import Ouroboros.Consensus.Block.Abstract
 import Ouroboros.Consensus.Ledger.Basics
+import Ouroboros.Consensus.Ledger.HD
 import Ouroboros.Consensus.Ticked
 import Ouroboros.Consensus.Util
 
@@ -86,6 +98,105 @@ import Ouroboros.Consensus.Util
 -- inputs exist and are still unspent, a block carries a sufficient
 -- cryptographic signature, etc.
 data family Validated x :: Type
+
+{-------------------------------------------------------------------------------
+  Tick a ledger state
+-------------------------------------------------------------------------------}
+
+-- | Whether we tell the ledger layer to compute ledger events
+--
+-- At the moment events are not emitted in any case in the consensus
+-- layer (i.e. there is no handler for those events, nor are they
+-- traced), so they are not really forced, we always discard
+-- them. This behavior does not incur big costs thanks to laziness.
+--
+-- By passing 'OmitLedgerEvents' we tell the Ledger layer to not even
+-- allocate thunks for those events, as we explicitly don't want them.
+data ComputeLedgerEvents = ComputeLedgerEvents | OmitLedgerEvents
+  deriving (Eq, Show, Generic, NoThunks)
+
+type IsLedger :: (Type -> Type) -> Type -> Constraint
+class
+  ( -- Requirements on the ledger state itself
+    Eq (l blk)
+  , NoThunks (l blk)
+  , Show (l blk)
+  , -- Requirements on 'LedgerCfg'
+    NoThunks (LedgerCfg l blk)
+  , -- Requirements on 'LedgerErr'
+    Show (LedgerErr l blk)
+  , Eq (LedgerErr l blk)
+  , NoThunks (LedgerErr l blk)
+  , -- Get the tip
+    --
+    -- See comment for 'applyChainTickLedgerResult' about the tip of the
+    -- ticked ledger.
+    GetTip (l blk)
+  , GetTip (Ticked l blk)
+  , -- The block axis of UTxO-HD: provides the opaque @'TickAndBlockDiff' blk@
+    -- that ticking and block application produces.
+    BlockSupportsLedgerHD blk
+  ) =>
+  IsLedger l blk
+  where
+  -- | Errors that can arise when updating the ledger
+  --
+  -- This is defined here rather than in 'ApplyBlock', since the /type/ of
+  -- these errors does not depend on the type of the block.
+  type LedgerErr l blk :: Type
+
+  -- | Apply "slot based" state transformations
+  --
+  -- When a block is applied to the ledger state, a number of things happen
+  -- purely based on the slot number of that block. For example:
+  --
+  -- * In Byron, scheduled updates are applied, and the update system state is
+  --   updated.
+  -- * In Shelley, delegation state is updated (on epoch boundaries).
+  --
+  -- The consensus layer must be able to apply such a "chain tick" function,
+  -- primarily when validating transactions in the mempool (which, conceptually,
+  -- live in "some block in the future") or when extracting valid transactions
+  -- from the mempool to insert into a new block to be produced.
+  --
+  -- This is not allowed to throw any errors. After all, if this could fail,
+  -- it would mean a /previous/ block set up the ledger state in such a way
+  -- that as soon as a certain slot was reached, /any/ block would be invalid.
+  --
+  -- Ticking a ledger state needs no values from the on-disk tables (in
+  -- particular it does not read the UTxO), but it may nonetheless /produce/ a
+  -- @'TickDiff' blk@ out of nothing: era transitions happen when ticking, and some
+  -- of them delete entries from the UTxO (e.g. the AVVM addresses removed at
+  -- the Shelley-to-Allegra boundary). That diff is returned alongside the
+  -- ticked state and must be composed with the block-application diff (and
+  -- forwarded onto any values read against the pre-tick state).
+  --
+  -- PRECONDITION: The slot number must be strictly greater than the slot at
+  -- the tip of the ledger (except for EBBs, obviously..).
+  --
+  -- NOTE: 'applyChainTickLedgerResult' should /not/ change the tip of the
+  -- underlying ledger state, which should still refer to the most recent
+  -- applied /block/. In other words, we should have:
+  --
+  -- prop> ledgerTipPoint (fst (applyChainTick cfg slot st)) == ledgerTipPoint st
+  applyChainTickLedgerResult ::
+    ComputeLedgerEvents ->
+    LedgerCfg l blk ->
+    SlotNo ->
+    l blk ->
+    LedgerResult blk (Ticked l blk, TickDiff blk)
+
+-- | 'lrResult' after 'applyChainTickLedgerResult'
+applyChainTick ::
+  IsLedger l blk =>
+  ComputeLedgerEvents ->
+  LedgerCfg l blk ->
+  SlotNo ->
+  l blk ->
+  (Ticked l blk, TickDiff blk)
+applyChainTick = lrResult ...: applyChainTickLedgerResult
+
+type LedgerError blk = LedgerErr LedgerState blk
 
 {-------------------------------------------------------------------------------
   Apply block to ledger state
