+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Node/ProtocolInfo.hs
@@ -11,8 +11,8 @@ import Data.Word
 import NoThunks.Class (NoThunks)
 import Ouroboros.Consensus.Block
 import Ouroboros.Consensus.Config
-import Ouroboros.Consensus.Ledger.Basics (Values)
 import Ouroboros.Consensus.Ledger.Extended
+import Ouroboros.Consensus.Ledger.HD (Values)
 import Ouroboros.Consensus.NodeId
 
 {-------------------------------------------------------------------------------
